Implemented 'fieldObject' trait.
